Port backbone app.
Adds the collections, views, and models that are needed to add sections.
The ported code removes the notion of multiple stages as there is only
one builder in this theme. Additionally, all TinyMCE and sortables
functionality is removed as it will likely be implemented in a different
way to decouple the section functionality from the base functionality.