Fix disappearing text column during sort.
When sorting text columns, one of the columns was disappearing. It turns
out that the nth-child selector was causing the disappearance. When
dragging the item, the nth-child selector begins to apply to different
columns, causing the wrong selectors to show/hide. By labelling the
columns according to which one it is, the proper show/hide status can be
maintained regardless of sorting.

Ref: #3