Init slide TinyMCEs only once.
Because the creation of a new section triggers initEditors() and the
slide creation also triggers initEditors(), the first slide in the
banner section could have the TinyMCE editor init'd multiple times
depending on a race condition. This results in multiple quicktag rows
being set up. Setting a flag to suppress the init function when the
section loads prevents the duel init.

Ref: #43