Add sidebar templates.
It is best if we try to keep the sidebar templating as standard as
possible. There are important hooks in get_sidebar() that plugins might
depend on. It is important to use that function to include them. This
will make it easier for child themers to customize the theme by
providing the sidebar templates that can be copied into the child theme.
Additionally, the boilerplate of wrapping HTML is fairly minimal so this
change should not add significant maintenance overhead.