Wrap oEmbeds in embed shortcode tag.
Because `wpautop()` must run on each content section individually, yet
the oEmbeds should not execute until they are displayed on the front
end, oEmbed URLs tend to get wrapped in <p> tags, which cause them to
not embed on the front end. This commit fakes the oEmbed when the
content is saved, but instead of rendering the oEmbed HTML, it just
wraps the URL in the embed shortcode tags (i.e., [embed]url[/embed]).
This strategy allows the URL to continue to be embedded correctly, even
when it is wrapped in <p> tags.

Fixes #71