Adjust sidebar ids and classes
Rather than altering the container id when the sidebar id is filtered,
add the sidebar id as another class. This way, styles that are specific
to the location of the sidebar will still be applied even if a non-default
sidebar is used in that location, as long as the styles use the id.