Allow greater precision for custom logo padding.
The custom logo's height is held by setting the padding for the <a> tag
that wraps the logo. The padding value is a width:height ratio. When
this value is calculated, it is cast to an absolute integer, which
causes it to lose precision. The net effect is that the image dimensions
can change and cause image distortion.

This commit attempts to fix the issue by allowing a greater level of
precision while still sanitizing the value.