Enable child themes to override some stylesheets and scripts
This commit introduces a new function, ttfmake_theme_file_url(), which
acts as a wrapper for locate_template(), taking the resulting path from
that function and converting it into the appropriate URL for inclusion
on the front end.

This new function is then applied to several of the auxiliary
stylesheets and scripts that are loaded in ttfmake_scripts().